refactored: dashboard.py split into multiple core files and gui files (src/gui/dashboard_views) + application.py split into multiple gui views (src/gui/app_views)
